requirements.txt
Twisted[tls]
service-identity
ldap3
PyYAML
